+++ src/main/java/me/tongfei/progressbar/ProgressBar.java
@@ -38,7 +38,7 @@ public class ProgressBar implements AutoCloseable {
         this(
                 task, initialMax, 1000, false, false,
                 System.err, ProgressBarStyle.COLORFUL_UNICODE_BLOCK,
-                "", 1, false, null,
+                "", 1L, false, null,
                 ChronoUnit.SECONDS, 0L, Duration.ZERO
         );
     }
@@ -62,7 +62,7 @@ public class ProgressBar implements AutoCloseable {
             boolean continuousUpdate,
             boolean clearDisplayOnFinish,
             PrintStream os,
-            DrawStyle style,
+            ProgressBarStyle style,
             String unitName,
             long unitSize,
             boolean showSpeed,
@@ -79,38 +79,6 @@ public class ProgressBar implements AutoCloseable {
                 ),
                 createConsoleConsumer(os)
         );
-    }
-
-    /**
-     * Creates a progress bar with the specific taskName name, initial maximum value,
-     * customized update interval (default 1000 ms), the PrintStream to be used, and output style.
-     * @param task Task name
-     * @param initialMax Initial maximum value
-     * @param updateIntervalMillis Update interval (default value 1000 ms)
-     * @param continuousUpdate Rerender every time the update interval happens regardless of progress count.
-     * @param style Output style (default value ProgressBarStyle.UNICODE_BLOCK)
-     * @param showSpeed Should the calculated speed be displayed
-     * @param speedFormat Speed number format
-     * @deprecated Use {@link ProgressBarBuilder} instead.
-     */
-    public ProgressBar(
-            String task,
-            long initialMax,
-            int updateIntervalMillis,
-            boolean continuousUpdate,
-            boolean clearDisplayOnFinish,
-            PrintStream os,
-            ProgressBarStyle style,
-            String unitName,
-            long unitSize,
-            boolean showSpeed,
-            DecimalFormat speedFormat,
-            ChronoUnit speedUnit,
-            long processed,
-            Duration elapsed
-    ) {
-        this(task, initialMax, updateIntervalMillis, continuousUpdate, clearDisplayOnFinish, os,
-        DrawStyle.from(style), unitName, unitSize, showSpeed, speedFormat, speedUnit, processed, elapsed);
     }
 
     /**
@@ -510,6 +478,11 @@ public class ProgressBar implements AutoCloseable {
     public static <T> Stream<T> wrap(T[] array, ProgressBarBuilder pbb) {
         pbb.setInitialMax(array.length);
         return wrap(Arrays.stream(array), pbb);
+    }
+
+    /** Creates a new builder to customize a progress bar. */
+    public static ProgressBarBuilder builder() {
+        return new ProgressBarBuilder();
     }
 
 }